docs(zbc_backend): add Doxygen docs, vtable conventions, improve types
Improve zbc_backend.h header with @file/@brief, detailed vtable comment
explaining implementation and return conventions. Add per-function Doxygen
docs. Update flen to intmax_t for larger files, heapinfo to uintptr_t for
portability. Use specific ZBC_ERR_* codes in riff functions.

(cherry picked from commit ef7003c6513e29b749fe6d40ca21f446491d2258)

+++ 3rdparty/semihost/src/zbc_backend_dummy.c
@@ -48,7 +48,7 @@ static int dummy_seek(void *ctx, int fd, int pos)
     return 0;
 }
 
-static int dummy_flen(void *ctx, int fd)
+static intmax_t dummy_flen(void *ctx, int fd)
 {
     (void)ctx;
     (void)fd;
@@ -167,8 +167,8 @@ static int dummy_get_cmdline(void *ctx, char *buf, size_t buf_size)
 }
 
 static int dummy_heapinfo(void *ctx,
-                          unsigned int *heap_base, unsigned int *heap_limit,
-                          unsigned int *stack_base, unsigned int *stack_limit)
+                          uintptr_t *heap_base, uintptr_t *heap_limit,
+                          uintptr_t *stack_base, uintptr_t *stack_limit)
 {
     (void)ctx;
     *heap_base = 0;
